Project/spark/requirements.txt
pyspark==3.5.3
google-cloud-storage==2.18.2
